[2023/12/26] JPA EntityManager Dao Example

+++ src/main/java/dataAccess/dto/Member.java
@@ -4,6 +4,6 @@ public class Member {
 
     int id;
     String name;
-    int point;
+    double point;
 
 }
